Reorganize files and namespaces
