+++ src/App/Vivianne.cli/Commands/Viv/VivCommand_Info.cs
@@ -3,7 +3,7 @@ using TheXDS.MCART.Helpers;
 using TheXDS.Vivianne.Models;
 using TheXDS.Vivianne.Serializers;
 
-namespace TheXDS.Vivianne.Commands;
+namespace TheXDS.Vivianne.Commands.Viv;
 
 /// <summary>
 /// Defines a command that allows the user to interact with a VIV file.
