+++ src/App/Vivianne.cli/Commands/Viv/VivCommand_Read.cs
@@ -1,7 +1,7 @@
 using System.CommandLine;
 using TheXDS.MCART.Types.Extensions;
 
-namespace TheXDS.Vivianne.Commands;
+namespace TheXDS.Vivianne.Commands.Viv;
 
 /// <summary>
 /// Defines a command that allows the user to interact with a VIV file.
